Refactor code and minor bug fixes
Since the whole app is the calculator, having a calculator object seemed
a bit unnecessary, making the code somewhat heavy to read with all those
".this". I removed the calculator object.

I increased the display's height to prevent showing the scrollbar
prematurely on Firefox.

Fixed a bug that caused 'NaN' being displayed when inserting a point
right after an operator.